+++ packages/core/src/field/setFieldInput/setFieldInput.ts
@@ -88,7 +88,12 @@ function setNestedInput(
     }
 
     // Set input for each array item
-    for (let index = 0; index < items.length; index++) {
+    for (
+      let index = 0;
+      // @ts-expect-error
+      index < arrayInput.length;
+      index++
+    ) {
       // Recursively set nested input
       setNestedInput(
         internalFieldStore.children[index],